+++ src/conf.rs
@@ -1,7 +1,7 @@
 //pub const STACK_BOTTOM = 0xbfff_fff0;
 //pub const TEXT_START = 0x0001_0000;
-pub const STACK_BOTTOM: u32 = 0x000F_4240;
-//pub const TEXT_START: u32 = 0x0000_03E8;
-pub const TEXT_START: u32 = 0x0000_0000;
+pub const STACK_BOTTOM: u64 = 0x000F_4240;
+//pub const TEXT_START: u64 = 0x0000_03E8;
+pub const TEXT_START: u64 = 0x0000_0000;
 
 pub const MEMORY_SIZE: usize = 100_000_000;
